+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.2] - 2024-02-08
+
+### Fixed
+
+- Fix glyph start and end when using `shape-run-cache`
+
 ## [0.11.1] - 2024-02-08
 
 ### Added